feat: testes avançados e validações de segurança no CRUD de produtos

+++ app/core/models/produto.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, DECIMAL
 from app.adapters.db.database import Base
 
 class Produto(Base):
@@ -7,4 +7,4 @@ class Produto(Base):
     id = Column(Integer, primary_key=True, index=True)
     nome = Column(String, nullable=False)
     descricao = Column(String, nullable=True)
-    preco = Column(Integer, nullable=False)
+    preco = Column(DECIMAL(10, 2), nullable=False)
